requirements.txt
# Core dependencies
gradio>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
tavily-python>=0.3.0

# Optional dependencies for better performance
uvicorn>=0.23.0
fastapi>=0.104.0